requirements.txt
django==4.2.1
djangorestframework
django-filter
Pillow
gunicorn
psycopg2-binary==2.8.5
beautifulsoup4==4.8.2
requests
numpy